Add watching for invite code to fire hook call
